Link click handler cleanup
Moved it out of header.js since it's not particular to the header. Also made it match on the href starting with a slash instead of on a class.
