+++ client/header.html
@@ -6,15 +6,15 @@
 </template>
 
 <template name="header_breadcrumb_chat">
-  <a href="/chat/{{type}}/{{id}}" class="chat-link{{#if inThisRoom}} current{{/if}}" title="Chat"><i class="fas fa-comments"></i></a>
+  <a href="/chat/{{type}}/{{id}}" class="{{#if inThisRoom}}current{{/if}}" title="Chat"><i class="fas fa-comments"></i></a>
 </template>
 
 <template name="header_breadcrumb_blackboard">
   <li class="{{#if active}}active{{/if}}">
-    <a href="/" class="home-link"><i class="fas fa-chalkboard"></i><span class="bb-omit-when-narrow"> Blackboard      
+    <a href="/"><i class="fas fa-chalkboard"></i><span class="bb-omit-when-narrow"> Blackboard      
     <i class="fas fa-th-list {{#if currentPageEquals "blackboard"}}current{{/if}}"></i></span></a>
     {{#if expertMode}}
-      <a href="/logistics" title="Logistics" class="logistics-link"><i class="fas fa-dolly {{#if currentPageEquals "logistics_page"}}current{{/if}}"></i></a>
+      <a href="/logistics" title="Logistics"><i class="fas fa-dolly {{#if currentPageEquals "logistics_page"}}current{{/if}}"></i></a>
     {{/if}}
     {{> header_breadcrumb_chat}}
   </li>
@@ -44,7 +44,7 @@
 
 <template name="header_breadcrumb_round">
   <li class="{{#if active}}active{{/if}}">
-    <a href="/rounds/{{id}}" class="rounds-link">
+    <a href="/rounds/{{id}}">
       <i class="fas fa-folder"></i><span class="bb-omit-when-narrow"> Round:</span> <span class="bb-abbrev-when-narrow" data-full="{{round.name}}" data-abbrev="{{abbrev round.name}}"></span>
     </a>
     {{> header_breadcrumb_chat}}
@@ -53,7 +53,7 @@
 
 <template name="header_breadcrumb_one_meta">
   <li class="{{#if active}}active{{/if}}">
-    <a href="/puzzles/{{id}}" class="puzzles-link">
+    <a href="/puzzles/{{id}}">
       <i class="fas fa-folder-open"></i><span class="bb-omit-when-narrow"> Meta:</span> <span class="bb-abbrev-when-narrow" data-full="{{puzzle.name}}" data-abbrev="{{abbrev puzzle.name}}"></span>
     </a>
     {{> header_breadcrumb_extra_links puzzle}}
@@ -94,7 +94,7 @@
 
 <template name="header_breadcrumb_puzzle">
   <li class="{{#if active}}active{{/if}}">
-    <a href="/puzzles/{{id}}" class="puzzles-link">
+    <a href="/puzzles/{{id}}">
       <i class="fas fa-book"></i><span class="bb-omit-when-narrow"> Puzzle:</span> <span class="bb-puzzle-name bb-abbrev-when-narrow" data-full="{{puzzle.name}}" data-abbrev="{{abbrev puzzle.name}}"></span>
     </a>
     {{> header_breadcrumb_extra_links puzzle}}
@@ -103,7 +103,7 @@
 
 <template name="header_breadcrumb_oplog">
   <li class="{{#if currentPageEquals "oplog"}}active{{/if}}">
-    <a href="/oplogs" class="oplogs-link">
+    <a href="/oplogs">
       <i class="fas fa-clock"></i> Operation Logs
     </a>
   </li>
@@ -111,7 +111,7 @@
 
 <template name="header_breadcrumb_facts">
   <li class="{{#if currentPageEquals "facts"}}active{{/if}}">
-    <a href="/facts" class="facts-link">
+    <a href="/facts">
       <i class="fas fa-wrench"></i> Server Facts
     </a>
   </li>
