+++ client/link.html
@@ -4,7 +4,7 @@
   {{else if this.editing}}
     {{text}}
   {{else}}
-    <a href="{{#if chat}}/chat{{/if}}/{{target.type}}/{{id}}" class="{{#if chat}}chat{{else}}{{target.type}}{{/if}}-link {{class}}" title="{{title}}">
+    <a href="{{#if chat}}/chat{{/if}}/{{target.type}}/{{id}}" title="{{title}}">
       {{#if icon}}
         <i class="{{icon}}" title="{{#unless title}}{{text}}{{/unless}}"></i>
       {{else}}
