+++ client/page.html
@@ -47,7 +47,7 @@
       {{else}}
         {{> embedded_chat}}
         <div class="bb-chat-pop-out {{#if not topRight}}in-margin{{/if}}">
-          <a href="/chat/{{type}}/{{id}}" class="chat-link bb-pop-out"
+          <a href="/chat/{{type}}/{{id}}" class="bb-pop-out"
               target="chat0"><i class="fas fa-external-link-alt"></i>
         Pop out</a>
         </div>
