+++ lib/model.js
@@ -1407,7 +1407,7 @@ if (Meteor.isServer) {
           : "";
       body = `has requested help: ${UI._escape(
         rawhow
-      )} (puzzle <a class=\"puzzles-link\" href=\"${objUrl}\">${UI._escape(
+      )} (puzzle <a href=\"${objUrl}\">${UI._escape(
         obj.name
       )}</a>${solverTimePart})`;
       Meteor.call("newMessage", {
